Revert "Merge branch 'gabrhr' into ericko"
This reverts commit a92b226d85a3e65325a274422f18512ccc0840da, reversing
changes made to 2401bf8f7b0d6a86f8879eaeda7408404beb319f.
