+++ src/components/controls/Controls.js
@@ -11,7 +11,6 @@ import DreamTitle from "./DreamTitle";
 import Divider from "./Divider"
 import RowCheckBox from "./RowCheckBox"
 import RangeTimePicker from "./RangeTimePicker"
-import InfoHelper from "./InfoHelper";
 
 export const Controls = {
     Input,
@@ -26,6 +25,5 @@ export const Controls = {
     Divider,
     DreamTitle,
     RowCheckBox,
-    RangeTimePicker, 
-    InfoHelper
+    RangeTimePicker
 }
